refactor: moves authentication APIs to a separate module
